# service/openapi.yaml
openapi: 3.1.0
info:
  title: Zailom WA Service
  version: 1.0.0
  summary: Central WhatsApp / Evolution API service — single source of truth for Zailom products.
  description: |
    All endpoints below (except `/healthz`, `/readyz`, `/v1/hooks/*`, and `/v1/admin/*`) require
    a tenant API key in the `Authorization: Bearer <key>` header.

    Admin endpoints require `X-Admin-Token: <ADMIN_TOKEN>`.

    Inbound Evolution webhooks (`/v1/hooks/evolution/:evoName`) are public but
    protected by an unguessable path and optionally by `X-Evolution-Token`.
  contact: { name: Zailom Platform }
servers:
  - url: https://wa.zailom.com
    description: Production
  - url: http://localhost:8080
    description: Local dev
...
security:
  - TenantKey: []
components:
  securitySchemes:
    TenantKey:
      type: http
      scheme: bearer
      bearerFormat: "zwa_live_<prefix>_<secret>"
    AdminToken:
      type: apiKey
      in: header
      name: X-Admin-Token
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:    { type: string, example: instance_not_found }
            message: { type: string, example: "instance not found" }
            details: { type: object, additionalProperties: true, nullable: true }
...
    Instance:
      type: object
      properties:
        id:                       { type: string, format: uuid }
        tenant_id:                { type: string, format: uuid }
        name:                     { type: string }
        evolution_instance_name:  { type: string }
        status:                   { type: string, enum: [pending, connecting, connected, disconnected, deleted] }
        connected_number:         { type: string, nullable: true }
        qr_code:                  { type: string, nullable: true, description: "base64 PNG data URI" }
        pairing_code:             { type: string, nullable: true }
        webhook_url:              { type: string, nullable: true }
        webhook_events:           { type: array, items: { type: string } }
        created_at:               { type: string, format: date-time }
        updated_at:               { type: string, format: date-time }
    NewInstance:
      type: object
      required: [name]
      properties:
        name:               { type: string, example: "atendimento-1" }
        webhook_url:        { type: string, format: uri, nullable: true }
        webhook_events:     { type: array, items: { type: string }, example: [MESSAGES_UPSERT, CONNECTION_UPDATE] }
        webhook_by_events:  { type: boolean, default: false }
        webhook_base64:     { type: boolean, default: false }
        config:             { type: object, additionalProperties: true }
...
  responses:
    Unauthorized:
      description: Invalid or missing API key
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    NotFound:
      description: Resource not found
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Validation:
      description: Payload validation failed
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    Upstream:
      description: Evolution API error
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
paths:
...
  /v1/instances/all-instances:
    get:
      tags: [Instances]
      summary: List all instances of the calling tenant
      responses:
        "200":
          description: OK
          content: { application/json: { schema: { type: object, properties: { data: { type: array, items: { $ref: "#/components/schemas/Instance" } } } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/instances/create:
    post:
      tags: [Instances]
      summary: Create instance (Evolution + local record)
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/NewInstance" } } }
      responses:
        "201":
          description: Created
          content:
            application/json:
              example:
                id: "018f...-uuid"
                evolution_instance_name: "booking-abc12345-atendimento-1-9f2a"
                qr_code: "data:image/png;base64,iVBORw0KG..."
                status: connecting
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/Validation" }
        "502": { $ref: "#/components/responses/Upstream" }
  /v1/instances/{id}:
    get:
      tags: [Instances]
      summary: Instance details
      parameters: [ { in: path, name: id, required: true, schema: { type: string, format: uuid } } ]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Instance" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/instances/{id}/connect:
    post:
      tags: [Instances]
      summary: Generate QR / pairing code
      parameters: [ { in: path, name: id, required: true, schema: { type: string, format: uuid } } ]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: { number: { type: string, description: "Optional; enables pairing-code flow" } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              example: { qr_code: "data:image/png;base64,...", pairing_code: null, expires_in: 60 }
  /v1/instance/connectionState/{id}:
    get:
      tags: [Instances]
      summary: Raw Evolution connection state
      parameters: [ { in: path, name: id, required: true, schema: { type: string, format: uuid } } ]
      responses: { "200": { description: Evolution response (state, wuid...) } }
...
  /v1/instances/{id}/refresh-status:
    post:
      tags: [Instances]
      summary: Sync local status with Evolution
      parameters: [ { in: path, name: id, required: true, schema: { type: string, format: uuid } } ]
      responses: { "200": { description: OK } }
  /v1/instances/{id}/restart:
    post:
      tags: [Instances]
      summary: Restart instance
      parameters: [ { in: path, name: id, required: true, schema: { type: string, format: uuid } } ]
      responses: { "200": { description: OK } }
  /v1/instances/{id}/logout:
    post:
      tags: [Instances]
      summary: Logout (keep the record)
      parameters: [ { in: path, name: id, required: true, schema: { type: string, format: uuid } } ]
      responses: { "200": { description: OK } }
  /v1/instances/{id}/delete:
    delete:
      tags: [Instances]
      summary: Logout + delete from Evolution + soft-delete locally
      parameters: [ { in: path, name: id, required: true, schema: { type: string, format: uuid } } ]
      responses: { "204": { description: Deleted } }
  /v1/instances/{id}/settings:
    get:
      tags: [Instances]
      summary: Get settings (rejectCall, groupsIgnore, alwaysOnline, ...)
      parameters: [ { in: path, name: id, required: true, schema: { type: string, format: uuid } } ]
      responses: { "200": { description: OK } }
  /v1/instances/{id}/settings/set:
    post:
      tags: [Instances]
      summary: Update settings
      parameters: [ { in: path, name: id, required: true, schema: { type: string, format: uuid } } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                rejectCall:      { type: boolean }
                msgCall:         { type: string }
                groupsIgnore:    { type: boolean }
                alwaysOnline:    { type: boolean }
                readMessages:    { type: boolean }
                readStatus:      { type: boolean }
                syncFullHistory: { type: boolean }
      responses: { "200": { description: OK } }

  # --------------------------------------------------------------- Messages
...
  /v1/instances/{id}/webhook/find:
    get:
      tags: [Webhooks (config)]
      summary: Get outbound webhook config (this service -> tenant)
      parameters: [ { in: path, name: id, required: true, schema: { type: string, format: uuid } } ]
      responses:
        "200":
          description: OK
          content:
            application/json:
              example:
                url: "https://booking.zailom.com/api/webhooks/wa"
                byEvents: false
                base64: false
                events: [MESSAGES_UPSERT, CONNECTION_UPDATE]
  /v1/instances/{id}/webhook/set:
    post:
      tags: [Webhooks (config)]
      summary: Configure outbound webhook
      parameters: [ { in: path, name: id, required: true, schema: { type: string, format: uuid } } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url:      { type: string, format: uri, nullable: true }
                byEvents: { type: boolean }
                base64:   { type: boolean }
                events:   { type: array, items: { type: string } }
      responses: { "200": { description: OK } }

  # --------------------------------------------------------------- Inbound (public)
...